server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
llama-index==0.10.12
llama-index-llms-groq==0.1.3
llama-index-embeddings-huggingface==0.1.4
llama-index-vector-stores-pinecone==0.1.3
pinecone-client==3.0.3
python-dotenv==1.0.0
GitPython==3.1.41
pydantic==2.5.3
